Rather than stop on the first permutation that reaches 0, let the algorithm to generate more paths up to the number of available pack sizes
The added test case covers an issue where 8x 3 1x 31 3x 151 was the first
candidate discovered, but there's a solution that uses fewer packs.

This allows us to maintain performance at large quantities where it's very
computationally expensive to traverse a graph of every possibility.

My reasoning here is that there must be diminishing returns on allowing more
permutations to be generated in the hope of finding a shorter path to 0.
Since quantities with larger packs are calculated first, it's likely that
any alternative paths will be due to slight differences in configuration
at the smallest pack sizes. So chances are that the shortest path has
already been generated after a certain number of attempts.