Generate smaller graphs when permutating very large quantities with small pack sizes
Allowing an exhaustive search to be performed from the initial quantity
results in 2-6gb memory usage for the added test cases. Calculating a
threshold which we subtract the largest pack sizes to without generating
a graph allows us to leave enough headroom in generating the necessary
permutations to find an exact fit.